LPS-66799 Refactor base DDM processor to a composite approach as it fits more with OSGi
