+++ modules/apps/collaboration/document-library/document-library-service/src/main/java/com/liferay/document/library/internal/exportimport/content/processor/FileEntryDDMFormValuesExportImportContentProcessor.java
@@ -14,11 +14,14 @@
 
 package com.liferay.document.library.internal.exportimport.content.processor;
 
-import com.liferay.dynamic.data.mapping.exportimport.content.processor.BaseDDMFormValuesExportImportContentProcessor;
+import com.liferay.dynamic.data.mapping.exportimport.content.processor.DDMFormValuesExportImportContentProcessor;
+import com.liferay.dynamic.data.mapping.storage.DDMFormValues;
 import com.liferay.exportimport.content.processor.ExportImportContentProcessor;
+import com.liferay.exportimport.kernel.lar.PortletDataContext;
 import com.liferay.portal.kernel.repository.model.FileEntry;
 
 import org.osgi.service.component.annotations.Component;
+import org.osgi.service.component.annotations.Reference;
 
 /**
  * @author Mate Thurzo
@@ -31,5 +34,42 @@ import org.osgi.service.component.annotations.Component;
 	service = {ExportImportContentProcessor.class}
 )
 public class FileEntryDDMFormValuesExportImportContentProcessor
-	extends BaseDDMFormValuesExportImportContentProcessor<FileEntry> {
+	implements ExportImportContentProcessor<FileEntry, DDMFormValues> {
+
+	@Override
+	public DDMFormValues replaceExportContentReferences(
+			PortletDataContext portletDataContext, FileEntry stagedModel,
+			DDMFormValues content, boolean exportReferencedContent,
+			boolean escapeContent)
+		throws Exception {
+
+		return ddmFormValuesExportImportContentProcessor.
+			replaceExportContentReferences(
+				portletDataContext, stagedModel, content,
+				exportReferencedContent, escapeContent);
+	}
+
+	@Override
+	public DDMFormValues replaceImportContentReferences(
+			PortletDataContext portletDataContext, FileEntry stagedModel,
+			DDMFormValues content)
+		throws Exception {
+
+		return ddmFormValuesExportImportContentProcessor.
+			replaceImportContentReferences(
+				portletDataContext, stagedModel, content);
+	}
+
+	@Override
+	public boolean validateContentReferences(
+		long groupId, DDMFormValues content) {
+
+		return ddmFormValuesExportImportContentProcessor.
+			validateContentReferences(groupId, content);
+	}
+
+	@Reference
+	protected DDMFormValuesExportImportContentProcessor
+		ddmFormValuesExportImportContentProcessor;
+
 }
