+++ modules/apps/forms-and-workflow/dynamic-data-lists/dynamic-data-lists-service/src/main/java/com/liferay/dynamic/data/lists/exportimport/content/processor/DDLRecordDDMFormValuesExportImportContentProcessor.java
@@ -15,10 +15,13 @@
 package com.liferay.dynamic.data.lists.exportimport.content.processor;
 
 import com.liferay.dynamic.data.lists.model.DDLRecord;
-import com.liferay.dynamic.data.mapping.exportimport.content.processor.BaseDDMFormValuesExportImportContentProcessor;
+import com.liferay.dynamic.data.mapping.exportimport.content.processor.DDMFormValuesExportImportContentProcessor;
+import com.liferay.dynamic.data.mapping.storage.DDMFormValues;
 import com.liferay.exportimport.content.processor.ExportImportContentProcessor;
+import com.liferay.exportimport.kernel.lar.PortletDataContext;
 
 import org.osgi.service.component.annotations.Component;
+import org.osgi.service.component.annotations.Reference;
 
 /**
  * @author Mate Thurzo
@@ -30,5 +33,39 @@ import org.osgi.service.component.annotations.Component;
 	service = {ExportImportContentProcessor.class}
 )
 public class DDLRecordDDMFormValuesExportImportContentProcessor
-	extends BaseDDMFormValuesExportImportContentProcessor<DDLRecord> {
+	implements ExportImportContentProcessor<DDLRecord, DDMFormValues> {
+
+	public DDMFormValues replaceExportContentReferences(
+			PortletDataContext portletDataContext, DDLRecord stagedModel,
+			DDMFormValues content, boolean exportReferencedContent,
+			boolean escapeContent)
+		throws Exception {
+
+		return ddmFormValuesExportImportContentProcessor.
+			replaceExportContentReferences(
+				portletDataContext, stagedModel, content,
+				exportReferencedContent, escapeContent);
+	}
+
+	public DDMFormValues replaceImportContentReferences(
+			PortletDataContext portletDataContext, DDLRecord stagedModel,
+			DDMFormValues content)
+		throws Exception {
+
+		return ddmFormValuesExportImportContentProcessor.
+			replaceImportContentReferences(
+				portletDataContext, stagedModel, content);
+	}
+
+	public boolean validateContentReferences(
+		long groupId, DDMFormValues content) {
+
+		return ddmFormValuesExportImportContentProcessor.
+			validateContentReferences(groupId, content);
+	}
+
+	@Reference
+	protected DDMFormValuesExportImportContentProcessor
+		ddmFormValuesExportImportContentProcessor;
+
 }
